requirements.txt
absl-py==1.4.0
addict==2.4.0
audioread==3.0.0
basicsr==1.4.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
facexlib==0.2.5
filterpy==1.4.5
fonttools==4.39.3
future==0.18.3
gfpgan==1.3.8
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.7.0
llvmlite==0.31.0
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
networkx==3.1
numba==0.48.0
numpy==1.23.0
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.5.5.64
packaging==23.0
Pillow==9.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
six==1.16.0
soundfile==0.12.1
tb-nightly==2.13.0a20230412
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.3.21
tqdm==4.45.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yapf==0.32.0
zipp==3.15.0

